WebSockets/tests/submissions/Opera/cookies/004.html

<!doctype html>
<title>WebSockets: setting HttpOnly cookies in ws response, checking document.cookie</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>

var timeout = setTimeout(function() {
  deleteCookie(false, 'timed out');
}, 10000);
var iframe = document.createElement('iframe');
try {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie_http');
  ws.onopen = function(e) {
    
    ws.onclose = debug;
    ws.close();
    deleteCookie(!(/ws_test=test/.test(document.cookie)), document.cookie);
  }
  ws.onclose = t.step_func(function() {assert_unreached()});
} catch(e) {
  assert_unreached(e);
}

function deleteCookie(testPassed, msg) {
  if (arguments.callee.done)
    return;
  arguments.callee.done = true;
  clearTimeout(timeout);
  setTestTimeout(10000);
  // remove cookie
  iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
  iframe.onload = function() {
    assert_equals(testPassed, true, msg);
    end();
  }
}
document.body.appendChild(iframe);
</script>
